Separate out the partitions into separate consumers.
The criteria for this system:

1) Give something which looks like a sarama.PartitionConsumer to allow
portability for existing code to use sarama-cluster with inimal
shimming.

2) Gracefully handle rebalances of partitions / discovering more
partitions by yielding the partitions on a channel as they are
found/added.

3) Maintain a go-like pattern; let the user of the library decide what
to do with each partition, with regards to concurrency, etc... we
don't run goroutines for them, and don't force them to respond to
callbacks on an unknown goroutine.