notify consumers of rebalance in progress #98
the problem with rebalance currently is that downstream consumers have no idea when it’s happening because the `Message()` channel will just block, but they are expected to commit offsets within `Synchronisation.DwellTime`

the workaround is to have a periodic timer and always commit at `Synchronisation.DwellTime / 2` interval, but that’s not ideal for either library or the downstream consumers.

the solution here is to use the `Notifications()` channel to notify downstream consumers of a rebalance in progress, so they can start committing current state, and also has a `MarkConsumersDone()` function that allows downstream consumers to notify sarama-cluster that the clients are ready for rebalancing, instead of waiting for the full `Synchronisation.DwellTime` to pass.