Use GroupProtocol chosen by CG Coordinator
The CG coordinator chooses a group protocol by inspecting the
ordered protocols in the JoinGroup requests and choosing one that
all members support.  The CG leader is expected to use this choice
to assign partitions.  Currently, though, the protocol on the CG
leader is read from config.  This isn't an issue if all consumers
are using sarama, but if consumers are using mixed libraries, then
it could result in the sarama CG leader choosing a protocol that
is not supported by other consumers.